requirements.txt
# Core dependencies
torch>=2.0.0
transformers>=4.30.0
peft>=0.4.0
accelerate>=0.20.0
bitsandbytes>=0.40.0
datasets>=2.12.0
evaluate>=0.4.0
scikit-learn>=1.2.2

# Data processing
dvc>=2.45.1
pypdf2>=3.0.0
python-docx>=0.8.11
pandas>=2.0.0
numpy>=1.24.0
tqdm>=4.65.0

# API and deployment
fastapi>=0.95.0
uvicorn>=0.22.0
pydantic>=1.10.0
python-multipart>=0.0.6
aiofiles>=23.1.0

# Evaluation
ragas>=0.0.16
nltk>=3.8.1

# Development and testing
pytest>=7.3.1
black>=23.3.0
isort>=5.12.0
mypy>=1.3.0
jupyter>=1.0.0

# Documentation
mkdocs>=1.4.3
mkdocs-material>=9.1.15 